requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.11.1
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
dotenv-python==0.0.1
et-xmlfile==1.1.0 ; python_version >= '3.6'
flask==2.2.2
future==0.18.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
graphviz==0.20.1
idna==3.4 ; python_version >= '3.5'
importlib-metadata==5.1.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
line-bot-sdk==1.14.0
markupsafe==2.1.1 ; python_version >= '3.7'
numpy==1.23.5 ; python_version < '3.10'
opencc==1.1.1
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1 ; python_version >= '3.6'
transitions==0.9.0
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.2 ; python_version >= '3.7'
zipp==3.11.0 ; python_version >= '3.7'
